Update Rumprun to latest Solo5
Update the build-rr.sh script to get the solo5 bindings object files
from the correct directory.
Update the ldscript for solo5 platforms
Update the solo5 git submodule.

AN important note is that Solo5 has introduced support for TLS and it
would be a good idea to integrate this support in Rumprun's scheduling.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>